fortress.pl: Updated logic to capture the full number of conns per-IP
In the previous logic, as soon as we reach the number of concurrent
connections from an IP, we block it. But this way, we do not know
how many were the total connections from that IP.

With this new approach, we move the check after we have already
parsed the file.
This way, we log the actual number of conns that the IP had.